fixes qualification section bug also
